fix: typescript upgrade and next video bundler
