+++ package.json
@@ -65,7 +65,7 @@
     "prettier": "^3.0",
     "prettier-plugin-tailwindcss": "^0.5.0",
     "tailwindcss": "^3.3.6",
-    "typescript": "^4.9.5"
+    "typescript": "5.6"
   },
   "lint-staged": {
     "**/*.{js,jsx,ts,tsx}": [
