+++ src/components/PromoVideo.tsx
@@ -1,34 +1,15 @@
 'use client';
 
-import { useInView } from 'framer-motion';
-import { useEffect, useRef } from 'react';
+import Video from 'next-video';
+
+import VideoAutoPlayer from '@/components/VideoAutoPlayer';
+
+import promoVideo from '/videos/dinox-promo.mp4';
 
 export default function PromoVideo() {
-  const ref = useRef<HTMLVideoElement>(null);
-  const isInView = useInView(ref);
-
-  useEffect(() => {
-    if (!isInView) {
-      ref.current?.pause();
-    } else {
-      ref.current?.play();
-    }
-  }, [isInView]);
-
   return (
     <div>
-      <video
-        ref={ref}
-        autoPlay={false}
-        muted={false}
-        controls={true}
-        loop={true}
-      >
-        <source
-          src='https://bkpvl280bbkyjsre.public.blob.vercel-storage.com/dinox-promo-azDIAnT87op5FThQu8yaKWuVrD9wyR.mp4'
-          type='video/mp4'
-        />
-      </video>
+      <Video as={VideoAutoPlayer} src={promoVideo}></Video>
     </div>
   );
 }
