+++ src/constant/env.ts
@@ -3,4 +3,4 @@ export const isLocal = process.env.NODE_ENV === 'development';
 
 export const showLogger = isLocal
   ? true
-  : process.env.NEXT_PUBLIC_SHOW_LOGGER === 'true' ?? false;
+  : process.env.NEXT_PUBLIC_SHOW_LOGGER === 'true';
