+++ src/globals.d.ts
@@ -3,3 +3,4 @@ declare module '*.png';
 declare module '*.jpg';
 declare module '*.jpeg';
 declare module '*.svg';
+declare module '*.mp4';
